Add function update_comparison_expression
The variable last_comparison_expr contains the last comparison expr
made. This expression is not added to the z3f expression because we
need to negate it later. Calling the function with an expression
as parameter, it will add the last expression to the z3f expression
and save the parameter in the variable.