[lab3]: Create function to get the branch to target
The function implements a depth-first or a breadth-first search
depending on the value of the boolean variable
`fuzzer.DEPTH_FIRST_SEARCH`.

It also implements a mechanism to put a target back to the bottom
of the queue if too many attempts to reach the target were made.
This is useful to avoid getting stuck on a branch that is not
reachable.

Related: #29